Preparing to make bulk delete internal only.
No active interest in this, so going to merge it into internal branches.
This work here is to minimise merge conflict as much as is possible.
