+++ core/src/test/java/org/apache/iceberg/hadoop/TestHadoopFileIO.java
@@ -18,50 +18,35 @@
  */
 package org.apache.iceberg.hadoop;
 
-import static org.apache.iceberg.hadoop.HadoopFileIO.BULK_DELETE_ENABLED;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.Mockito.mockStatic;
 
 import java.io.File;
 import java.io.IOException;
 import java.io.UncheckedIOException;
 import java.nio.file.Files;
-import java.util.AbstractMap;
-import java.util.Collection;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 import java.util.Random;
-import java.util.Set;
 import java.util.UUID;
 import java.util.Vector;
 import java.util.stream.Collectors;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.fs.BulkDelete;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.security.UserGroupInformation;
-import org.apache.iceberg.CatalogUtil;
 import org.apache.iceberg.TestHelpers;
 import org.apache.iceberg.common.DynMethods;
-import org.apache.iceberg.exceptions.NotFoundException;
 import org.apache.iceberg.io.BulkDeletionFailureException;
 import org.apache.iceberg.io.FileIO;
 import org.apache.iceberg.io.FileIOParser;
-import org.apache.iceberg.io.FileInfo;
 import org.apache.iceberg.io.ResolvingFileIO;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableMap;
 import org.apache.iceberg.relocated.com.google.common.collect.Lists;
-import org.apache.iceberg.relocated.com.google.common.collect.Sets;
 import org.apache.iceberg.relocated.com.google.common.collect.Streams;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
-import org.mockito.MockedStatic;
 
 public class TestHadoopFileIO {
   private final Random random = new Random(1);
@@ -73,32 +58,14 @@ public class TestHadoopFileIO {
 
   @BeforeEach
   public void before() throws Exception {
-    resetFileIOBinding(false);
-  }
-
-  /**
-   * Resets fs and hadoopFileIO fields to a configuration built from the supplied settings. The two
-   * settings are not orthogonal; if bulk delete is enabled then deleteFiles() hands off to the FS
-   * and its bulk delete operation, while single file delete may still go through trash.
-   *
-   * @param bulkDelete use bulk delete
-   * @throws UncheckedIOException on failures to create a new FS.
-   */
-  private void resetFileIOBinding(boolean bulkDelete) {
     Configuration conf = new Configuration();
-    conf.setBoolean(BULK_DELETE_ENABLED, bulkDelete);
-    try {
-      FileSystem.closeAllForUGI(UserGroupInformation.getCurrentUser());
-      fs = FileSystem.getLocal(conf);
-    } catch (IOException e) {
-      throw new UncheckedIOException(e);
-    }
-    hadoopFileIO = new HadoopFileIO(fs.getConf());
+    fs = FileSystem.getLocal(conf);
+
+    hadoopFileIO = new HadoopFileIO(conf);
   }
 
   @Test
-  public void testListPrefixAndDeleteFiles() {
-    resetFileIOBinding(true);
+  public void testListPrefix() {
     Path parent = new Path(tempDir.toURI());
 
     List<Integer> scaleSizes = Lists.newArrayList(1, 1000, 2500);
@@ -115,15 +82,8 @@ public class TestHadoopFileIO {
             });
 
     long totalFiles = scaleSizes.stream().mapToLong(Integer::longValue).sum();
-    final String parentString = parent.toUri().toString();
-    final List<FileInfo> files = Streams.stream(hadoopFileIO.listPrefix(parentString)).toList();
-    assertThat(files.size())
-        .describedAs("Files found under %s", parentString)
+    assertThat(Streams.stream(hadoopFileIO.listPrefix(parent.toUri().toString())).count())
         .isEqualTo(totalFiles);
-
-    // Delete the files.
-    final Iterator<String> locations = files.stream().map(FileInfo::location).iterator();
-    hadoopFileIO.deleteFiles(() -> locations);
   }
 
   @Test
@@ -133,13 +93,9 @@ public class TestHadoopFileIO {
     fs.createNewFile(randomFilePath);
 
     // check existence of the created file
-    final String path = randomFilePath.toUri().toString();
-    assertThat(hadoopFileIO.newInputFile(path).exists()).isTrue();
+    assertThat(hadoopFileIO.newInputFile(randomFilePath.toUri().toString()).exists()).isTrue();
     fs.delete(randomFilePath, false);
-    assertThat(hadoopFileIO.newInputFile(path).exists()).isFalse();
-    assertThatThrownBy(() -> hadoopFileIO.newInputFile(path).getLength())
-        .isInstanceOf(NotFoundException.class)
-        .hasMessageContaining(path);
+    assertThat(hadoopFileIO.newInputFile(randomFilePath.toUri().toString()).exists()).isFalse();
   }
 
   @Test
@@ -170,218 +126,23 @@ public class TestHadoopFileIO {
         .hasMessageContaining("java.io.FileNotFoundException");
   }
 
-  @ParameterizedTest
-  @ValueSource(booleans = {true, false})
-  public void testDeleteFiles(boolean bulkDelete) {
-    resetFileIOBinding(bulkDelete);
+  @Test
+  public void testDeleteFiles() {
     Path parent = new Path(tempDir.toURI());
     List<Path> filesCreated = createRandomFiles(parent, 10);
-    hadoopFileIO.deleteFiles(filesCreated.stream().map(Path::toString).toList());
-    assertPathsDoNotExist(filesCreated);
+    hadoopFileIO.deleteFiles(
+        filesCreated.stream().map(Path::toString).collect(Collectors.toList()));
+    filesCreated.forEach(
+        file -> assertThat(hadoopFileIO.newInputFile(file.toString()).exists()).isFalse());
   }
 
-  /** Test both delete paths through CatalogUtil. */
-  @ParameterizedTest
-  @ValueSource(booleans = {true, false})
-  public void testBulkDeleteThroughCatalogUtil(boolean bulkDelete) {
-    resetFileIOBinding(bulkDelete);
-    List<Path> filesCreated = createRandomFiles(new Path(tempDir.toURI()), 10);
-    CatalogUtil.deleteFiles(
-        hadoopFileIO, filesCreated.stream().map(Path::toString).toList(), "uncommitted");
-    assertPathsDoNotExist(filesCreated);
-  }
-
-  /**
-   * Mock the bulk delete API as unavailable, verify that CatalogUtil.deleteFiles() doesn't raise an
-   * exception -simply failsl to delete the files. A warning is logged; this isn't picked up in the
-   * test.
-   */
   @Test
-  public void testCatalogUtilWhenBulkDeleteUnavailable() {
-    resetFileIOBinding(true);
-    try (MockedStatic<BulkDeleter> mocked = mockStatic(BulkDeleter.class)) {
-      mocked.when(BulkDeleter::apiAvailable).thenReturn(false);
-      List<Path> filesCreated = createRandomFiles(new Path(tempDir.toURI()), 10);
-      CatalogUtil.deleteFiles(
-          hadoopFileIO, filesCreated.stream().map(Path::toString).toList(), "uncommitted");
-      assertPathsExist(filesCreated);
-    }
-  }
-
-  @ParameterizedTest
-  @ValueSource(booleans = {true, false})
-  public void testDeleteFilesErrorHandling(boolean bulkDelete) {
-    resetFileIOBinding(bulkDelete);
-    hadoopFileIO = new HadoopFileIO(fs.getConf());
-    assertThat(hadoopFileIO.useBulkDeleteApi())
-        .describedAs("Bulk Delete API use")
-        .isEqualTo(bulkDelete);
-    Path parent = new Path(tempDir.toURI());
-
-    List<String> filesToDelete =
+  public void testDeleteFilesErrorHandling() {
+    List<String> filesCreated =
         random.ints(2).mapToObj(x -> "fakefsnotreal://file-" + x).collect(Collectors.toList());
-    // one file in the local FS which doesn't actually exist but whose scheme is valid
-    // this MUST NOT be recorded as a failure
-    final String localButMissing = new Path(parent, "file-not-exist").toUri().toString();
-    filesToDelete.add(localButMissing);
-    final String exists = touch(new Path(parent, "exists")).toString();
-    filesToDelete.add(exists);
-    assertThatThrownBy(() -> hadoopFileIO.deleteFiles(filesToDelete))
-        .describedAs("Exception raised by deleteFiles()")
+    assertThatThrownBy(() -> hadoopFileIO.deleteFiles(filesCreated))
         .isInstanceOf(BulkDeletionFailureException.class)
-        .hasMessage("Failed to delete 2 files")
-        .matches(
-            (e) -> ((BulkDeletionFailureException) e).numberFailedObjects() == 2,
-            "Wrong number of failures");
-    assertFileDoesNotExist(localButMissing);
-    assertFileDoesNotExist(exists);
-  }
-
-  /**
-   * Mock the bulk delete API as unavailable, then expect a delete with bulk delete enabled to fail.
-   */
-  @Test
-  public void testDeleteFailureWhenBulkDeleteUnavailable() {
-    resetFileIOBinding(true);
-    try (MockedStatic<BulkDeleter> mocked = mockStatic(BulkDeleter.class)) {
-      mocked.when(BulkDeleter::apiAvailable).thenReturn(false);
-      assertThatThrownBy(() -> hadoopFileIO.deleteFiles(List.of(tempDir.toURI().toString())))
-          .isInstanceOf(IllegalStateException.class)
-          .hasMessageContaining(BULK_DELETE_ENABLED);
-    }
-  }
-
-  /**
-   * Create a dir with a child, expect a failure, but also expect a file named in the same list
-   * provided to have been deleted.
-   */
-  @Test
-  public void testBulkDeleteNonEmptyDirectory() {
-    resetFileIOBinding(true);
-    Path parent = new Path(tempDir.toURI());
-    final String file = touch(new Path(parent, "file")).toString();
-    final File directory = new File(tempDir, "dir");
-    directory.mkdirs();
-    touch(new Path(parent, "dir/child"));
-
-    final String dirString = directory.toURI().toString();
-    assertFileExists(dirString);
-    // directory delete resulted in the whole operation reported as a failure.
-    assertThatThrownBy(() -> hadoopFileIO.deleteFiles(List.of(dirString, file)))
-        .isInstanceOf(BulkDeletionFailureException.class)
-        .hasMessageContaining("Failed to delete 1 files");
-
-    // directory is still there
-    assertFileExists(dirString);
-    // the file has been deleted
-    assertFileDoesNotExist(file);
-  }
-
-  /**
-   * With bulk delete disabled, deleteFiles() still works when bulk delete has not been enabled.
-   * This shows the classic invocation path is active.
-   */
-  @Test
-  public void testDeleteFilesWithBulkDeleteNotFound() {
-    try (MockedStatic<BulkDeleter> mocked = mockStatic(BulkDeleter.class)) {
-      mocked.when(BulkDeleter::apiAvailable).thenReturn(false);
-      Path parent = new Path(tempDir.toURI());
-      List<Path> filesCreated = createRandomFiles(parent, 5);
-      hadoopFileIO.deleteFiles(filesCreated.stream().map(Path::toString).toList());
-      assertPathsDoNotExist(filesCreated);
-    }
-  }
-
-  /**
-   * A stub BulkDelete implementation allows for a page size greater than 1 to be tested, through
-   * the DeleteContext class along with reporting of partial failures.
-   */
-  @Test
-  public void testBulkDeleteLargerContext() {
-
-    final Path root = new Path("/");
-
-    // stub delete with page size of two, path #3 will be rejected on a deleteBatch() call
-    final StubBulkDelete stubBulkDelete = new StubBulkDelete(2, 3);
-    final BulkDeleter.DeleteContext context = new BulkDeleter.DeleteContext(root, stubBulkDelete);
-    assertThat(context.pageSize()).isEqualTo(2);
-    assertThat(context.size()).isEqualTo(0);
-    assertThat(context.pageSizeReached()).isFalse();
-    // add one path, assert state changes
-    final Path p1 = new Path(root, "p1");
-    context.add(p1);
-    assertThat(context.size()).isEqualTo(1);
-    assertThat(context.pageSizeReached()).isFalse();
-    // add a second path, the page is now complete
-    final Path p2 = new Path(root, "p2");
-    context.add(p2);
-    assertThat(context.pageSizeReached()).isTrue();
-    // take a snapshot, it has the expected size and the context is reset
-    final Set<Path> snapshot = context.snapshotDeletedFiles();
-    assertThat(snapshot).hasSize(2);
-    assertThat(context.size()).isEqualTo(0);
-    assertThat(context.pageSizeReached()).isFalse();
-    // perform the delete
-    assertThat(context.deleteBatch(snapshot).failures()).isEmpty();
-    final Path p3 = new Path(root, "p3");
-    context.add(p3);
-    final Path p4 = new Path(root, "p4");
-    context.add(p4);
-    assertThat(context.pageSizeReached()).isTrue();
-    // delete all without a snapshot (as is done at the end of deleteFiles());
-    // this is going report a failure.
-    assertThat(context.deleteBatch(context.deletedFiles()).failures())
-        .hasSize(1)
-        .allSatisfy(
-            r -> {
-              assertThat(r.path()).isEqualTo(p3);
-              assertThat(r.errorText()).isEqualTo("simulated failure");
-            });
-    // still deleted everything
-    assertThat(stubBulkDelete.deleteCount).isEqualTo(4);
-    assertThat(stubBulkDelete.deletedFiles).containsExactlyInAnyOrder(p1, p2, p4);
-    context.close();
-  }
-
-  /** Stub implementation of BulkDelete to verify invocation and failure handling. */
-  private static final class StubBulkDelete implements BulkDelete {
-    private final int pageSize;
-    private final int failOnDeleteCount;
-    private final Set<Path> deletedFiles = Sets.newHashSet();
-    private int deleteCount;
-
-    StubBulkDelete(int pageSize, int failOnDeleteCount) {
-      this.pageSize = pageSize;
-      this.failOnDeleteCount = failOnDeleteCount;
-    }
-
-    @Override
-    public int pageSize() {
-      return pageSize;
-    }
-
-    @Override
-    public Path basePath() {
-      return new Path("/");
-    }
-
-    @Override
-    public List<Map.Entry<Path, String>> bulkDelete(Collection<Path> paths)
-        throws IOException, IllegalArgumentException {
-      List<Map.Entry<Path, String>> entries = Lists.newArrayList();
-      for (Path path : paths) {
-        deleteCount++;
-        if (deleteCount == failOnDeleteCount) {
-          entries.add(new AbstractMap.SimpleEntry<>(path, "simulated failure"));
-        } else {
-          deletedFiles.add(path);
-        }
-      }
-      return entries;
-    }
-
-    @Override
-    public void close() throws IOException {}
+        .hasMessage("Failed to delete 2 files");
   }
 
   @ParameterizedTest
@@ -459,62 +220,19 @@ public class TestHadoopFileIO {
 
   private List<Path> createRandomFiles(Path parent, int count) {
     Vector<Path> paths = new Vector<>();
-    random.ints(count).parallel().forEach(i -> paths.add(touch(new Path(parent, "file-" + i))));
+    random
+        .ints(count)
+        .parallel()
+        .forEach(
+            i -> {
+              try {
+                Path path = new Path(parent, "file-" + i);
+                paths.add(path);
+                fs.createNewFile(path);
+              } catch (IOException e) {
+                throw new UncheckedIOException(e);
+              }
+            });
     return paths;
-  }
-
-  /**
-   * Create a file at a path, overwriting any existing file.
-   *
-   * @param path path of file.
-   * @return the path.
-   */
-  private Path touch(Path path) {
-    try {
-      fs.create(path, true).close();
-      return path;
-    } catch (IOException e) {
-      throw new UncheckedIOException(e);
-    }
-  }
-
-  /**
-   * Assert a file exists.
-   *
-   * @param file URI to file/dir.
-   */
-  private void assertFileExists(String file) {
-    assertThat(hadoopFileIO.newInputFile(file).exists())
-        .describedAs("File %s must exist", file)
-        .isTrue();
-  }
-
-  /**
-   * Assert a file does not exist.
-   *
-   * @param file URI to file/dir.
-   */
-  private void assertFileDoesNotExist(String file) {
-    assertThat(hadoopFileIO.newInputFile(file).exists())
-        .describedAs("File %s must exist", file)
-        .isFalse();
-  }
-
-  /**
-   * Given a list of paths, assert none of them exist.
-   *
-   * @param paths list of paths
-   */
-  private void assertPathsDoNotExist(List<Path> paths) {
-    paths.forEach(path -> assertFileDoesNotExist(path.toString()));
-  }
-
-  /**
-   * Given a list of paths, assert all of them exist.
-   *
-   * @param paths list of paths
-   */
-  private void assertPathsExist(List<Path> paths) {
-    paths.forEach(path -> assertFileExists(path.toString()));
   }
 }
